test(cli): ✅ added unit tests for executeGitMessage
