+++ src/utils/executeCommand.test.ts
@@ -1,176 +1,12 @@
 /* eslint-disable jest/no-large-snapshots */
-
-import { executeDryRun, formatCommitMessage } from './executeCommand'
 import { defaultConfig } from '../defaults'
-
-const setupFormatCommitMessage = (config = {}, answers = {}) => {
-  return formatCommitMessage(
-    { ...defaultConfig, ...config },
-    {
-      body: 'this an amazing feature, lots of details',
-      breaking: 'breaks everything',
-      issues: '#123',
-      scope: '*',
-      subject: 'a cool new feature',
-      type: 'feat',
-      ...answers,
-    },
-    true
-  )
-}
-
-describe('formatCommitMessage', () => {
-  it('should format commit message with everything', () => {
-    const formattedMessage = setupFormatCommitMessage()
-
-    expect(formattedMessage).toMatchInlineSnapshot(`
-      "feat(*): 🎸 a cool new feature
-
-      this an amazing feature, lots of details
-
-      BREAKING CHANGE: 🧨 breaks everything
-
-      🏁 Closes: #123"
-    `)
-  })
-
-  it('should format commit message with no emojis', () => {
-    const formattedMessage = setupFormatCommitMessage({ disableEmoji: true })
-
-    expect(formattedMessage).toMatchInlineSnapshot(`
-      "feat(*): a cool new feature
-
-      this an amazing feature, lots of details
-
-      BREAKING CHANGE: breaks everything
-
-      Closes: #123"
-    `)
-  })
-
-  it('should format commit message with no body', () => {
-    const formattedMessage = setupFormatCommitMessage(defaultConfig, {
-      body: '',
-    })
-
-    expect(formattedMessage).toMatchInlineSnapshot(`
-      "feat(*): 🎸 a cool new feature
-
-      BREAKING CHANGE: 🧨 breaks everything
-
-      🏁 Closes: #123"
-    `)
-  })
-
-  it('should format commit message with multiline body', () => {
-    const formattedMessage = setupFormatCommitMessage(defaultConfig, {
-      body: '\n',
-    })
-
-    expect(formattedMessage).toMatchInlineSnapshot(`
-      "feat(*): 🎸 a cool new feature
-
-      BREAKING CHANGE: 🧨 breaks everything
-
-      🏁 Closes: #123"
-    `)
-  })
-
-  it('should format commit message with no scope', () => {
-    const formattedMessage = setupFormatCommitMessage(defaultConfig, {
-      scope: '',
-    })
-
-    expect(formattedMessage).toMatchInlineSnapshot(`
-      "feat: 🎸 a cool new feature
-
-      this an amazing feature, lots of details
-
-      BREAKING CHANGE: 🧨 breaks everything
-
-      🏁 Closes: #123"
-    `)
-  })
-
-  it('should format commit message with no issues', () => {
-    const formattedMessage = setupFormatCommitMessage(defaultConfig, {
-      issues: '',
-    })
-
-    expect(formattedMessage).toMatchInlineSnapshot(`
-      "feat(*): 🎸 a cool new feature
-
-      this an amazing feature, lots of details
-
-      BREAKING CHANGE: 🧨 breaks everything"
-    `)
-  })
-
-  it('should format commit message with no breaking change', () => {
-    const formattedMessage = setupFormatCommitMessage(defaultConfig, {
-      breaking: '',
-    })
-
-    expect(formattedMessage).toMatchInlineSnapshot(`
-      "feat(*): 🎸 a cool new feature
-
-      this an amazing feature, lots of details
-
-      🏁 Closes: #123"
-    `)
-  })
-
-  it('should wrap commit message', () => {
-    const formattedMessage = setupFormatCommitMessage(defaultConfig, {
-      body:
-        'volutpat commodo sed egestas egestas fringilla phasellus faucibus scelerisque eleifend donec',
-      breaking:
-        'volutpat commodo sed egestas egestas fringilla phasellus faucibus scelerisque eleifend donec pretium',
-    })
-
-    expect(formattedMessage).toMatchInlineSnapshot(`
-      "feat(*): 🎸 a cool new feature
-
-      volutpat commodo sed egestas egestas fringilla phasellus faucibus
-      scelerisque eleifend donec
-
-      BREAKING CHANGE: 🧨 volutpat commodo sed egestas egestas fringilla
-      phasellus faucibus scelerisque eleifend donec pretium
-
-      🏁 Closes: #123"
-    `)
-  })
-  it('should allow double quotes in message', () => {
-    const formattedMessage = setupFormatCommitMessage(defaultConfig, {
-      subject: 'this has "quotes"',
-      body: '',
-      breaking: '',
-      issues: '',
-    })
-
-    expect(formattedMessage).toMatchInlineSnapshot(
-      `"feat(*): 🎸 this has \\\\\\"quotes\\\\\\""`
-    )
-  })
-  it('should allow backticks quotes in message', () => {
-    const formattedMessage = setupFormatCommitMessage(defaultConfig, {
-      subject: 'this has `quotes`',
-      body: '',
-      breaking: '',
-      issues: '',
-    })
-
-    expect(formattedMessage).toMatchInlineSnapshot(
-      `"feat(*): 🎸 this has \\\\\`quotes\\\\\`"`
-    )
-  })
-})
+import * as fns from './executeCommand'
 
 describe('executeDryRun', () => {
   it('should console log git message', () => {
     const spy = jest.spyOn(console, 'log').mockImplementation(jest.fn())
 
-    executeDryRun('feat(cli): 🎸 initial release')
+    fns.executeDryRun('feat(cli): 🎸 initial release')
 
     expect(spy).toMatchInlineSnapshot(`
       [MockFunction] {
@@ -196,5 +32,29 @@ describe('executeDryRun', () => {
         ],
       }
     `)
+  })
+})
+const answers = {
+  body: 'this an amazing feature, lots of details',
+  breaking: 'breaks everything',
+  issues: '#123',
+  scope: '*',
+  subject: 'a cool new feature',
+  type: 'feat',
+}
+
+describe('executeGitMessage', () => {
+  it('should call executeCommand', () => {
+    const spy = jest.spyOn(fns, 'executeCommand').mockImplementation(jest.fn())
+
+    fns.executeGitMessage({ config: defaultConfig, answers }, {})
+    expect(spy).toHaveBeenCalledTimes(1)
+  })
+
+  it('should call executeDryRun', () => {
+    const spy = jest.spyOn(fns, 'executeDryRun').mockImplementation(jest.fn())
+
+    fns.executeGitMessage({ config: defaultConfig, answers }, { dryRun: true })
+    expect(spy).toHaveBeenCalledTimes(1)
   })
 })
